fix: added method text_over_input to generate fill-in boxes, adjusted method rename_files_recursively to rename subdirs, basedir and image files according to dates and channels

+++ batch_rename_files_in_folders.py
@@ -3,42 +3,58 @@ import re
 # import Path class from pathlib
 from pathlib import Path
 
-import PySimpleGUI as sg
+import PySimpleGUI as sG
 
 
 # import glob
-def rename_files_recursively(root_path, date1, date2, date3, date4, date5, date6, date7, date8, date9, date10):
+
+def text_over_input(text, input_size, dates_length):
+    return sG.Column(
+        [[sG.Text(text, pad=(0, 3))]] + [[sG.Input(key=text + str(counter), size=(input_size, 1), pad=(0, 3))] for
+                                         counter in range(dates_length)], pad=(5, 5))
+
+
+def rename_files_recursively(root_path, dapi, inputs):
     # change directory
     os.chdir(root_path)
     # display the current directory
     # cwd = os.getcwd()
     input_replacements = []
-    search_input_terms = []
-    dicti = {}
-    for i in range(10):
-        dicti[date1[i].split(" ")[0]] = date1[i].split(" ")[1:]
-    print(dicti)
-    channels = []
-    #    print(len(channels))
-    count = 0
-    for channel in channels:
-        if channel != "":
-            search_input_terms.append("c" + str(count))
-            input_replacements.append(channel)
-            # print(count)
-            # print(channel)
-        count += 1
+    search_input_terms = ["c0"]
+    input_replacements.append(dapi)
 
     standard_search_terms = ["stitching-Image Export", "-Stitching", "-Image Export", " - Copy",
-                             "-Background subtraction", "_alpha-SMA-555_CD177-647", "_ORG", "stitching", "-",
-                             " - Copy"]
-    standard_replacements = ["", "", "", "", "", "", "", "", "_", "_autobrightctrst"]
+                             "-Background subtraction", "_alpha-SMA-555_CD177-647", "_ORG", "stitching", "-"]
+    standard_replacements = ["", "", "", "", "", "", "", "", "_"]
 
     count = 0
     cwd = Path.cwd()
     # create new directory if does not exist
     # if not os.path.isdir(dest_dir):
     #    os.mkdir(dest_dir)
+    base_dir = os.path.basename(cwd)
+    new_base_dir_name = base_dir
+    # print(base_dir)
+    new_basepath = ""
+    for counter, term in enumerate(standard_search_terms):
+        if term in base_dir:
+            # print(base_dir)
+            new_base_dir_name = new_base_dir_name.replace(term, standard_replacements[counter])
+            # print(new_base_dir_name)
+            back_dir = os.path.dirname(os.path.dirname(cwd))
+            new_basepath = os.path.join(back_dir, new_base_dir_name)
+    if new_base_dir_name != base_dir and not os.path.exists(new_basepath):
+        os.rename(cwd, new_basepath)
+        # print(new_basepath)
+    for subdir in os.listdir(cwd):
+        new_subdir_name = subdir
+        for counter, term in enumerate(standard_search_terms):
+            if term in subdir:
+                # print(subdir)
+                new_subdir_name = new_subdir_name.replace(term, standard_replacements[counter])
+                # print(new_subdir_name)
+        if not os.path.exists(new_subdir_name):
+            os.rename(os.path.join(cwd, subdir), os.path.join(cwd, new_subdir_name))
     for dir_path, subdirs, file_names in os.walk(cwd):
         # for filenames in glob.iglob('Path.cwd() / '**' / '*.jpg', recursive=True):
         # print(file_names)
@@ -47,22 +63,38 @@ def rename_files_recursively(root_path, date1, date2, date3, date4, date5, date6
                 name, extension = os.path.splitext(filename)
                 substrings = name.split("_")
                 # print(substrings[0])
-                for i in range(len(substrings)):
-                    if re.match(".*c\d.*", substrings[i]):
-                        substrings[i] = substrings[i].replace(substrings[i],
-                                                              'c' + substrings[i][substrings[i].index('c') + 1])
-                        # print(substrings[i])
+                for counter in range(len(substrings)):
+                    if re.match(r'.*c\d.*', substrings[counter]):
+                        substrings[counter] = substrings[counter].replace(substrings[counter],
+                                                                          'c' + substrings[counter][
+                                                                              substrings[counter].index('c') + 1])
+                        # print(substrings[counter])
                 new_name = '_'.join(substrings)
+                for idate in inputs.keys():
+                    if idate != "" and idate in new_name:
+                        # print(dir_path)
+                        # print(new_name)
+                        if "c1" in name:
+                            new_name = new_name.replace("c1", inputs.get(idate)[0])
+                            # print(new_name)
+                        if "c2" in new_name:
+                            new_name = new_name.replace("c2", inputs.get(idate)[1])
+                        if "c3" in new_name:
+                            new_name = new_name.replace("c3", inputs.get(idate)[2])
                 # print(new_name)
                 search_terms = standard_search_terms + search_input_terms
                 replacements = standard_replacements + input_replacements
-                for i, term in enumerate(search_terms):
+                for counter, term in enumerate(search_terms):
                     if term in name:
-                        new_name = new_name.replace(term, replacements[i])
-                        continue
-                if name != new_name:
+                        new_name = new_name.replace(term, replacements[counter])
+                new_file_path = os.path.join(dir_path, new_name + extension)
+                if name != new_name and not os.path.exists(new_file_path):
+                    # print(new_file_path)
+                    # print(name)
+                    # print(new_name)
+                    # print(not os.path.exists(new_file_path))
                     os.rename(os.path.join(dir_path, filename),
-                              os.path.join(dir_path, new_name + extension))
+                              new_file_path)
                     count += 1
 
     print(f"{count} files were renamed recursively from root {cwd}")
@@ -73,37 +105,34 @@ if __name__ == "__main__":
     #        for file in files:
     #            if file.endswith(".txt") and file.__contains__("shading missing use automatic"):
     #                print(os.path.join(root, file))
-    sg.theme("DarkTeal2")
+    font = ('Courier New', 11)
+    # sg.theme('DarkBlue4')
+    sG.set_options(font=font)
+    size = 15
+    dates_number = 10
+    cols = (('dates', size), ('channel 1', size), ('channel 2', size), ('channel 3', size))
     layout = [
-        [sg.T("")],
-        [sg.Text("Choose a folder: "), sg.Input(key="-IN2-", change_submits=True, enable_events=True),
-         sg.FolderBrowse(key="-IN-")],
+        [sG.T("")],
+        [sG.Text("Choose a folder: "), sG.Input(key="-IN2-", change_submits=True, enable_events=True),
+         sG.FolderBrowse(key="-IN-")],
+        [sG.T("")],
         # [sg.Listbox(
         #   values=[], enable_events=True, size=(50, 5), key="-DatesChannels LIST-"
         # )
         # ],
-
-        [sg.Text("Channel0", size=(5, 1)), sg.InputText("0dapi", key="ch0")],
-        [sg.Text("date1", size=(5, 1)), sg.InputText(key="date1")],
-        [sg.Text("date2", size=(5, 1)), sg.InputText(key="date2")],
-        [sg.Text("date3", size=(5, 1)), sg.InputText(key="date3")],
-        [sg.Text("date4", size=(5, 1)), sg.InputText(key="date4")],
-        [sg.Text("date5", size=(5, 1)), sg.InputText(key="date5")],
-        [sg.Text("date6", size=(5, 1)), sg.InputText(key="date6")],
-        [sg.Text("date7", size=(5, 1)), sg.InputText(key="date7")],
-        [sg.Text("date8", size=(5, 1)), sg.InputText(key="date8")],
-        [sg.Text("date9", size=(5, 1)), sg.InputText(key="date9")],
-        [sg.Text("date10", size=(5, 1)), sg.InputText(key="date10")],
-        [sg.Button("Submit"), sg.Button("Cancel")],
+        [sG.Text("Dapi Channel", size=(15, 1)), sG.InputText("0dapi", key="ch0", size=(15, 1))],
+        [sG.T("")],
+        [*[text_over_input(*col, dates_number) for col in cols]],
+        [sG.Button("Submit"), sG.Button("Cancel")]
     ]
 
-    ###Building Window
-    window = sg.Window('My File Browser', layout, size=(600, 500))
+    # Building Window
+    window = sG.Window('My File Browser', layout, keep_on_top=True, element_justification='c')
 
     while True:
         event, values = window.read()
         # print(values["-IN2-"])
-        if event == sg.WIN_CLOSED or event == "Exit" or event == "Cancel":
+        if event == sG.WIN_CLOSED or event == "Exit" or event == "Cancel":
             break
         elif event == "-IN2-":
             folder = values["-IN2-"]
@@ -120,46 +149,68 @@ if __name__ == "__main__":
                 # and not f.__contains__(                    'shading missing use automatic')
             ]
 
-            myDict = {}
+            input_dates_channels = {}
             if len(fnames) == 1:
                 with open(os.path.join(folder, fnames[0])) as f:
                     lines = f.readlines()
                     for i, line in enumerate(lines):
-                        if re.match("^\d{6}$", line):
+                        if re.match(r'^\d{6}$', line):
                             channels = []
                             for j in range(i + 1, len(lines), 1):
                                 # print(lines[j])
-                                if re.match("^c\d.*\w*$", lines[j]):
+                                if re.match(r'^c\d.*\w*$', lines[j]):
                                     channels.append(lines[j].strip())
                                     # print(channels[0])
                                 else:
                                     break
-                                myDict[line.strip()] = channels
-                    #          print(myDict)
-            datech = []
-            for key in myDict.keys():
-                strin = key
-                stri = ""
-                for value in myDict.get(key):
-                    stri = stri + value.split(" ")[1] + " "
-                strin = strin + " " + stri + " "
-                datech.append(strin)
-            for i in range(len(datech)):
-                chan = "date" + str(i + 1)
-                window[chan].update(datech[i])
+                                input_dates_channels[line.strip()] = channels
+            date_channels_to_edit = []
+            for date in input_dates_channels.keys():
+                date_channels_string = date
+                channels = ""
+                for channel in input_dates_channels.get(date):
+                    # print(channel.split(" ")[1])
+                    channels = channels + channel.split(" ")[1] + " "
+                date_channels_string = date_channels_string + " " + channels.strip()
+                date_channels_to_edit.append(date_channels_string)
+
+            for i in range(len(date_channels_to_edit)):
+                idates = "dates" + str(i)
+                ichOne = "channel 1" + str(i)
+                ichTwo = "channel 2" + str(i)
+                ichThree = "channel 3" + str(i)
+                ivalues = date_channels_to_edit[i].split(" ")
+                if len(ivalues) == 4:
+                    window[idates].update(ivalues[0])
+                    window[ichOne].update(ivalues[1])
+                    window[ichTwo].update(ivalues[2])
+                    window[ichThree].update(ivalues[3])
+                if len(ivalues) == 3:
+                    window[idates].update(ivalues[0])
+                    window[ichOne].update(ivalues[1])
+                    window[ichTwo].update(ivalues[2])
+                if len(values) == 2:
+                    window[idates].update(ivalues[0])
+                    window[ichOne].update(ivalues[1])
+                if len(ivalues) == 1:
+                    window[idates].update(ivalues[0])
+                else:
+                    continue
+
         elif event == "Submit":
             if values["-IN2-"] == "":
-                sg.popup_error("Please choose a directory")
+                sG.popup_error("Please choose a directory")
             if values["ch0"] == "":
-                sg.popup_error("Please give at least channel 0")
+                sG.popup_error("Please give at least channel 0")
             try:
-                # print(values["-IN-"])
-                rename_files_recursively(values["-IN2-"], values["ch0"], values["date1"], values["date2"],
-                                         values["date3"], values["date4"], values["date5"], values["date6"],
-                                         values["date7"], values["date8"], values["date9"], values["date10"])
+                input_dates_channels_updated = {}
+                for i in range(dates_number):
+                    input_dates_channels_updated[values["dates" + str(i)]] = [values["channel 1" + str(i)],
+                                                                              values["channel 2" + str(i)],
+                                                                              values["channel 3" + str(i)]]
+                # print(input_dates_channels_updated)
+                rename_files_recursively(values["-IN2-"], values["ch0"], input_dates_channels_updated)
             except:
                 values["-IN2-"] = ""
                 values["ch0"] = ""
                 pass
-
-# TDo: cover the case if the same files as have to be saved are already in the folder
